NASA SWMF & OMNI magnetosphere data model integration, player attach phase visual update, Resonance field line improvement, enemy spawn indicators, team goal amplifier feedback, player health indicator test

+++ Assets/Scripts/Scoring/ScoreRuntimeTypes.cs
@@ -64,9 +64,9 @@ namespace Massive.Scoring
 
         // Personal multiplier. Retained under the original field name for
         // compatibility with existing scoring presentation and telemetry.
-        public int multiplier;
+        public double multiplier;
         public int teamAmplifierMultiplier;
-        public int combinedMultiplier;
+        public double combinedMultiplier;
         public long finalMilliElectronVolts;
         public long previousTeamTotalMilliElectronVolts;
         public long newTeamTotalMilliElectronVolts;
@@ -80,7 +80,7 @@ namespace Massive.Scoring
         public int playerID;
         public int teamID;
         public int acceptedAwardCount;
-        public int highestMultiplier;
+        public double highestMultiplier;
         public long baseMilliElectronVolts;
         public long finalMilliElectronVolts;
     }
